Move ecl_keyworddata_to_df() to common
